increase performance

+++ main.py
@@ -23,36 +23,36 @@ try:
             print("Reporter id : ", reporterid)
             image_list.append([caseid, licenseplatepic])
 
-        platenumber = None
-        # Get neural network prediction
-        for image in image_list:
-            # preprocess image
-            path = '{}.jpg'.format(image[0])
-            with open(path, 'wb') as file:
-                file.write(image[1])
-            platenumber = Start.recognize(path)
-            # delete file to save storage
-            os.remove(path)
+        if image_list != []:
+            platenumber = None
+            # Get neural network prediction
+            for image in image_list:
+                # preprocess image
+                path = '{}.jpg'.format(image[0])
+                with open(path, 'wb') as file:
+                    file.write(image[1])
+                platenumber = Start.recognize(path)
+                # delete file to save storage
+                os.remove(path)
 
-            platetuple = (platenumber,)
+                platetuple = (platenumber,)
 
-            # Get the row with vehicle number to get offender's id TABLE : vehicleregistration
-            cursor2 = connection.cursor()
-            query2 = ("SELECT id FROM vehicleregistration "
-                      "WHERE licenseplateno = %s")
+                # Get the row with vehicle number to get offender's id TABLE : vehicleregistration
+                cursor2 = connection.cursor()
+                query2 = ("SELECT id FROM vehicleregistration "
+                          "WHERE licenseplateno = %s")
 
-            cursor2.execute(query2, (platetuple))
-            for (id) in cursor2:
-                print("Case id : ", id)
-                criminalid = id[0]
+                cursor2.execute(query2, (platetuple))
+                for (id) in cursor2:
+                    print("Case id : ", id)
+                    criminalid = id[0]
 
-            # Update the offender's id TABLE : cases
-            cursor3 = connection.cursor()
-            query3 = """ UPDATE cases
-                    SET aadharid = %s
-                    WHERE caseid = %s """
-            cursor3.execute(query3, (criminalid, image[0]))
-
+                # Update the offender's id TABLE : cases
+                cursor3 = connection.cursor()
+                query3 = """ UPDATE cases
+                        SET aadharid = %s
+                        WHERE caseid = %s """
+                cursor3.execute(query3, (criminalid, image[0]))
 
 except Error as e:
     print("Error while connecting to MySQL", e)
